allow DataFrames with sequence instead of timestamps
