+++ dp_mobility_report/constants.py
@@ -43,6 +43,14 @@ ELEMENTS = OVERVIEW_ELEMENTS + PLACE_ELEMENTS + OD_ELEMENTS + USER_ELEMENTS
 TESSELLATION_ELEMENTS = (
     PLACE_ELEMENTS + OD_ELEMENTS + [USER_TILE_COUNT, MOBILITY_ENTROPY]
 )  # TODO: include or exclude travel_time and jump_length?
+TIMESTAMP_ANALYSES = [
+    TRIPS_OVER_TIME,
+    TRIPS_PER_WEEKDAY,
+    TRIPS_PER_HOUR,
+    VISITS_PER_TILE_TIMEWINDOW,
+    TRAVEL_TIME,
+    USER_TIME_DELTA,
+]
 
 SEGMENTS_AND_ELEMENTS = SEGMENTS + ELEMENTS
