+++ dp_mobility_report/model/preprocessing.py
@@ -1,12 +1,235 @@
 import logging
-from typing import Optional
+import warnings
+from typing import Any, List, Optional, Tuple, Union
 
 import geopandas as gpd
 import numpy as np
 import pandas as pd
+from geopandas import GeoDataFrame
+from pandas import DataFrame
 from shapely.geometry import Point
 
 from dp_mobility_report import constants as const
+
+
+def validate_inclusion_exclusion(
+    analysis_selection: Optional[List[str]], analysis_exclusion: Optional[List[str]]
+) -> Tuple[Optional[List[str]], Optional[List[str]]]:
+
+    if analysis_selection is not None and analysis_exclusion is not None:
+        warnings.warn(
+            "The parameter `analysis_exclusion' will be ignored, as `analysis_selection' is set as well."
+        )
+        analysis_exclusion = None
+
+    if analysis_selection == [const.ALL]:
+        warnings.warn(
+            "['all'] as input is deprecated. Use 'None' (default) instead to include all analyses."
+        )
+        analysis_selection = None
+
+    return analysis_selection, analysis_exclusion
+
+
+def validate_input(
+    df: DataFrame,
+    tessellation: Optional[GeoDataFrame],
+    privacy_budget: Optional[Union[int, float]],
+    max_trips_per_user: Optional[int],
+    analysis_selection: Optional[List[str]],
+    analysis_exclusion: Optional[List[str]],
+    budget_split: dict,
+    disable_progress_bar: bool,
+    evalu: bool,
+    user_privacy: bool,
+    timewindows: Union[List[int], np.ndarray],
+    max_travel_time: Optional[int],
+    bin_range_travel_time: Optional[int],
+    max_jump_length: Optional[Union[int, float]],
+    bin_range_jump_length: Optional[Union[int, float]],
+    max_radius_of_gyration: Optional[Union[int, float]],
+    bin_range_radius_of_gyration: Optional[Union[int, float]],
+    seed_sampling: Optional[int],
+) -> None:
+    if not isinstance(df, DataFrame):
+        raise TypeError("'df' is not a Pandas DataFrame.")
+
+    if tessellation is None:
+        warnings.warn(
+            "No tessellation has been specified. All analyses based on the tessallation will be omitted."
+        )
+
+    if not ((tessellation is None) or isinstance(tessellation, GeoDataFrame)):
+        raise TypeError("'tessellation' is not a Geopandas GeoDataFrame.")
+
+    if not ((max_trips_per_user is None) or isinstance(max_trips_per_user, int)):
+        raise TypeError("'max_trips_per_user' is not numeric.")
+    if (max_trips_per_user is not None) and (max_trips_per_user < 1):
+        raise ValueError("'max_trips_per_user' has to be greater 0.")
+
+    if analysis_selection is not None:
+        if not isinstance(analysis_selection, list):
+            raise TypeError("'analysis_selection' is not a list.")
+
+        if not set(analysis_selection).issubset(
+            const.SEGMENTS_AND_ELEMENTS + [const.ALL]
+        ):
+            raise ValueError(
+                f"Unknown analyses in {analysis_selection}. Only elements from {const.SEGMENTS_AND_ELEMENTS} are valid inputs."
+            )
+
+    if analysis_exclusion is not None:
+        if not isinstance(analysis_exclusion, list):
+            raise TypeError("'analysis_exclusion' is not a list.")
+        if not set(analysis_exclusion).issubset(const.SEGMENTS_AND_ELEMENTS):
+            raise ValueError(
+                f"Unknown analyses in {analysis_exclusion}. Only elements from {const.SEGMENTS_AND_ELEMENTS} are valid inputs."
+            )
+
+    if not isinstance(budget_split, dict):
+        raise TypeError("'budget_split' is not a dict.")
+    if not set(budget_split.keys()).issubset(const.ELEMENTS):
+        raise ValueError(
+            f"Unknown analyses in {budget_split}. Only elements from {const.ELEMENTS} are valid inputs as dictionary keys."
+        )
+    if not all(isinstance(x, int) for x in list(budget_split.values())):
+        raise ValueError(
+            f"Not all elements in 'budget_split' are integers: {list(budget_split.values())}."
+        )
+
+    if not isinstance(timewindows, (list, np.ndarray)):
+        raise TypeError("'timewindows' is not a list or a numpy array.")
+
+    timewindows = (
+        np.array(timewindows) if isinstance(timewindows, list) else timewindows
+    )
+    if not all([np.issubdtype(item, int) for item in timewindows]):
+        raise TypeError("not all items of 'timewindows' are integers.")
+
+    if privacy_budget is not None:
+        _validate_numeric_greater_zero(
+            privacy_budget, f"{privacy_budget=}".split("=")[0]
+        )
+    _validate_numeric_greater_zero(max_travel_time, f"{max_travel_time=}".split("=")[0])
+    _validate_numeric_greater_zero(
+        bin_range_travel_time, f"{bin_range_travel_time=}".split("=")[0]
+    )
+    _validate_numeric_greater_zero(max_jump_length, f"{max_jump_length=}".split("=")[0])
+    _validate_numeric_greater_zero(
+        bin_range_jump_length, f"{bin_range_jump_length=}".split("=")[0]
+    )
+    _validate_numeric_greater_zero(
+        max_radius_of_gyration, f"{max_radius_of_gyration=}".split("=")[0]
+    )
+    _validate_numeric_greater_zero(
+        bin_range_radius_of_gyration, f"{bin_range_radius_of_gyration=}".split("=")[0]
+    )
+    _validate_bool(user_privacy, f"{user_privacy=}".split("=")[0])
+    _validate_bool(evalu, f"{user_privacy=}".split("=")[0])
+    _validate_bool(disable_progress_bar, f"{user_privacy=}".split("=")[0])
+
+    if not ((seed_sampling is None) or isinstance(seed_sampling, int)):
+        raise TypeError("'seed_sampling' is not an integer.")
+    if (seed_sampling is not None) and (seed_sampling <= 0):
+        raise ValueError("'seed_sampling' has to be greater 0.")
+
+
+def _validate_numeric_greater_zero(var: Any, name: str) -> None:
+    if not ((var is None) or isinstance(var, (int, float))):
+        raise TypeError(f"{name} is not numeric.")
+    if (var is not None) and (var <= 0):
+        raise ValueError(f"'{name}' has to be greater 0.")
+
+
+def _validate_bool(var: Any, name: str) -> None:
+    if not isinstance(var, bool):
+        raise TypeError(f"'{name}' is not type boolean.")
+
+
+# unify different input options of analyses (segments and elements) to be included / excluded as excluded elements, i.e., 'analysis_exclusion'
+def clean_analysis_exclusion(
+    analysis_selection: Optional[List[str]],
+    analysis_exclusion: Optional[List[str]],
+    has_tessellation: bool,
+    has_timestamps: bool,
+) -> List[str]:
+    # TODO: without timestamp: add w/o timestamp analyses to exclude_analysis
+
+    def _remove_elements(elements: list, remove_list: list) -> list:
+        return [e for e in elements if e not in remove_list]
+
+    if analysis_selection is not None:
+        analysis_exclusion = const.ELEMENTS
+
+        # remove all elements of segments
+        if const.OVERVIEW in analysis_selection:
+            analysis_exclusion = _remove_elements(
+                analysis_exclusion, const.OVERVIEW_ELEMENTS
+            )
+        if const.PLACE_ANALYSIS in analysis_selection:
+            analysis_exclusion = _remove_elements(
+                analysis_exclusion, const.PLACE_ELEMENTS
+            )
+        if const.OD_ANALYSIS in analysis_selection:
+            analysis_exclusion = _remove_elements(analysis_exclusion, const.OD_ELEMENTS)
+        if const.USER_ANALYSIS in analysis_selection:
+            analysis_exclusion = _remove_elements(
+                analysis_exclusion, const.USER_ELEMENTS
+            )
+
+        # remove single elements
+        analysis_exclusion = _remove_elements(analysis_exclusion, analysis_selection)
+
+    elif analysis_exclusion is not None:
+        # deduplicate list in case there are duplicates as input (otherwise `remove` might fail)
+        analysis_exclusion = list(set(analysis_exclusion))
+
+        if const.OVERVIEW in analysis_exclusion:
+            analysis_exclusion += const.OVERVIEW_ELEMENTS
+            analysis_exclusion.remove(const.OVERVIEW)
+
+        if const.PLACE_ANALYSIS in analysis_exclusion:
+            analysis_exclusion += const.PLACE_ELEMENTS
+            analysis_exclusion.remove(const.PLACE_ANALYSIS)
+
+        if const.OD_ANALYSIS in analysis_exclusion:
+            analysis_exclusion += const.OD_ELEMENTS
+            analysis_exclusion.remove(const.OD_ANALYSIS)
+
+        if const.USER_ANALYSIS in analysis_exclusion:
+            analysis_exclusion += const.USER_ELEMENTS
+            analysis_exclusion.remove(const.USER_ANALYSIS)
+
+    else:
+        analysis_exclusion = []
+
+    if not has_tessellation:
+        analysis_exclusion += const.TESSELLATION_ELEMENTS
+
+    if not has_timestamps:
+        analysis_exclusion += const.TIMESTAMP_ANALYSES
+
+    # deduplicate in case analyses and segments were included
+    analysis_exclusion = list(set(analysis_exclusion))
+
+    # sort according to order
+    analysis_exclusion.sort(key=lambda i: const.ELEMENTS.index(i))
+
+    return analysis_exclusion
+
+
+def clean_budget_split(budget_split: dict, analysis_exclusion: List[str]) -> dict:
+    intersec = set(budget_split.keys()).intersection(analysis_exclusion)
+    if len(intersec) != 0:
+        warnings.warn(
+            f"A `budget_split`is specified for the analyses {intersec} even though they are excluded."
+            "As they will be excluded, the `budget_split` specification will be ignored for these analyses."
+        )
+
+    # remove all analyses that are excluded according to `analysis_exclusion``
+    remaining_analyses = set(budget_split.keys()) - set(analysis_exclusion)
+    budget_split = {analysis: budget_split[analysis] for analysis in remaining_analyses}
+    return budget_split
 
 
 def preprocess_tessellation(tessellation: gpd.GeoDataFrame) -> gpd.GeoDataFrame:
@@ -47,10 +270,15 @@ def _validate_columns(df: pd.DataFrame) -> pd.DataFrame:
         raise TypeError("Column 'lng' is not of type float.")
     if const.DATETIME not in df.columns:
         raise ValueError("Column 'datetime' must be present in data.")
-    try:
-        df.loc[:, const.DATETIME] = pd.to_datetime(df[const.DATETIME])
-    except Exception as ex:
-        raise TypeError("Column 'datetime' cannot be cast to datetime.") from ex
+    if pd.core.dtypes.common.is_integer_dtype(df[const.DATETIME]):
+        warnings.warn(
+            "Column 'datetime' consists of type int. It is only interpreted as sequence information and thus all time related analyses are omitted."
+        )
+    else:
+        try:
+            df.loc[:, const.DATETIME] = pd.to_datetime(df[const.DATETIME])
+        except Exception as ex:
+            raise TypeError("Column 'datetime' cannot be cast to datetime.") from ex
     return df
 
 
@@ -78,11 +306,13 @@ def preprocess_data(
         columns.append(const.TILE_ID)
     df = df.loc[:, columns]
 
-    # create time related variables
-    df.loc[:, const.HOUR] = df[const.DATETIME].dt.hour
-    df.loc[:, const.IS_WEEKEND] = np.select(
-        [df[const.DATETIME].dt.weekday > 4], ["weekend"], default="weekday"
-    )
+    # only create time realted variables if df has timestamps
+    if pd.core.dtypes.common.is_datetime64_dtype(df[const.DATETIME]):
+        # create time related variables
+        df.loc[:, const.HOUR] = df[const.DATETIME].dt.hour
+        df.loc[:, const.IS_WEEKEND] = np.select(
+            [df[const.DATETIME].dt.weekday > 4], ["weekend"], default="weekday"
+        )
 
     # remove waypoints
     df = df.sort_values(const.DATETIME).groupby(const.TID).nth([0, -1])
