+++ dp_mobility_report/report/html/html_templates/config_segment.html
@@ -3,4 +3,5 @@
 <p class="warning">
     {{privacy_info}}
     {{tessellation_info}}
+    {{timestamp_info}}
 </p>
